Added combo-box and functionality to plot data from previously stored dataframes
