+++ main.py
@@ -1,3 +1,4 @@
+from gui import startGui
 try:
     from scrape import storeScrape
     from scrape import scrape
@@ -5,6 +6,7 @@ try:
     import json
     import pandas as pd
     from io import StringIO
+    import os
 
 except ModuleNotFoundError:
     print('Please download dependencies from requirement.txt')
@@ -13,16 +15,3 @@ except Exception as ex:
 
 if __name__ == '__main__':
     startGui()
-    #print("test")
-    #twitter_handle, filepath = startGui()
-    #count = 200
-    #jsonResponse = scrape(twitter_handle, count)
-    #df = storeScrape(jsonResponse, filepath)
-    #plot(filepath,twitter_handle)
-    
-
-    
-    #filepath = r'C:\Users\rikar\Documents\Skola\KAU\Projekt\dataframe\{handle}.csv'.format(handle=twitter_handle.lower())
-    #jsonResponse = scrape(twitter_handle, count);
-    #df = storeScrape(jsonResponse, filepath)
-
